+++ src/main/java/com/studyolle/settings/form/Notifications.java
@@ -1,4 +1,4 @@
-package com.studyolle.settings;
+package com.studyolle.settings.form;
 
 import lombok.Data;
