+++ src/main/java/com/studyolle/settings/form/PasswordForm.java
@@ -1,6 +1,5 @@
-package com.studyolle.settings;
+package com.studyolle.settings.form;
 
-import com.studyolle.domain.Account;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
@@ -8,7 +7,7 @@ import org.hibernate.validator.constraints.Length;
 @NoArgsConstructor
 @Data
 public class PasswordForm {
-    
+
     @Length(min = 8, max = 50)
     private String newPassword;
