+++ src/main/java/com/studyolle/settings/form/Profile.java
@@ -1,4 +1,4 @@
-package com.studyolle.settings;
+package com.studyolle.settings.form;
 
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
